+++ server/models/raw/index.ts
@@ -1,71 +1,71 @@
-import PermissionsModel from '../models/Permissions';
+import PermissionsModel from '../../../app/models/server/models/Permissions';
 import { PermissionsRaw } from './Permissions';
-import RolesModel from '../models/Roles';
+import RolesModel from '../../../app/models/server/models/Roles';
 import { RolesRaw } from './Roles';
-import SubscriptionsModel from '../models/Subscriptions';
+import SubscriptionsModel from '../../../app/models/server/models/Subscriptions';
 import { SubscriptionsRaw } from './Subscriptions';
-import SettingsModel from '../models/Settings';
+import SettingsModel from '../../../app/models/server/models/Settings';
 import { SettingsRaw } from './Settings';
-import UsersModel from '../models/Users';
+import UsersModel from '../../../app/models/server/models/Users';
 import { UsersRaw } from './Users';
-import SessionsModel from '../models/Sessions';
+import SessionsModel from '../../../app/models/server/models/Sessions';
 import { SessionsRaw } from './Sessions';
-import RoomsModel from '../models/Rooms';
+import RoomsModel from '../../../app/models/server/models/Rooms';
 import { RoomsRaw } from './Rooms';
-import LivechatCustomFieldModel from '../models/LivechatCustomField';
+import LivechatCustomFieldModel from '../../../app/models/server/models/LivechatCustomField';
 import { LivechatCustomFieldRaw } from './LivechatCustomField';
-import LivechatTriggerModel from '../models/LivechatTrigger';
+import LivechatTriggerModel from '../../../app/models/server/models/LivechatTrigger';
 import { LivechatTriggerRaw } from './LivechatTrigger';
-import LivechatDepartmentModel from '../models/LivechatDepartment';
+import LivechatDepartmentModel from '../../../app/models/server/models/LivechatDepartment';
 import { LivechatDepartmentRaw } from './LivechatDepartment';
-import LivechatDepartmentAgentsModel from '../models/LivechatDepartmentAgents';
+import LivechatDepartmentAgentsModel from '../../../app/models/server/models/LivechatDepartmentAgents';
 import { LivechatDepartmentAgentsRaw } from './LivechatDepartmentAgents';
-import LivechatRoomsModel from '../models/LivechatRooms';
+import LivechatRoomsModel from '../../../app/models/server/models/LivechatRooms';
 import { LivechatRoomsRaw } from './LivechatRooms';
-import MessagesModel from '../models/Messages';
+import MessagesModel from '../../../app/models/server/models/Messages';
 import { MessagesRaw } from './Messages';
-import LivechatExternalMessagesModel from '../models/LivechatExternalMessages';
+import LivechatExternalMessagesModel from '../../../app/models/server/models/LivechatExternalMessages';
 import { LivechatExternalMessageRaw } from './LivechatExternalMessages';
-import LivechatVisitorsModel from '../models/LivechatVisitors';
+import LivechatVisitorsModel from '../../../app/models/server/models/LivechatVisitors';
 import { LivechatVisitorsRaw } from './LivechatVisitors';
-import LivechatInquiryModel from '../models/LivechatInquiry';
+import LivechatInquiryModel from '../../../app/models/server/models/LivechatInquiry';
 import { LivechatInquiryRaw } from './LivechatInquiry';
-import IntegrationsModel from '../models/Integrations';
+import IntegrationsModel from '../../../app/models/server/models/Integrations';
 import { IntegrationsRaw } from './Integrations';
-import EmojiCustomModel from '../models/EmojiCustom';
+import EmojiCustomModel from '../../../app/models/server/models/EmojiCustom';
 import { EmojiCustomRaw } from './EmojiCustom';
-import WebdavAccountsModel from '../models/WebdavAccounts';
+import WebdavAccountsModel from '../../../app/models/server/models/WebdavAccounts';
 import { WebdavAccountsRaw } from './WebdavAccounts';
-import OAuthAppsModel from '../models/OAuthApps';
+import OAuthAppsModel from '../../../app/models/server/models/OAuthApps';
 import { OAuthAppsRaw } from './OAuthApps';
-import CustomSoundsModel from '../models/CustomSounds';
+import CustomSoundsModel from '../../../app/models/server/models/CustomSounds';
 import { CustomSoundsRaw } from './CustomSounds';
-import CustomUserStatusModel from '../models/CustomUserStatus';
+import CustomUserStatusModel from '../../../app/models/server/models/CustomUserStatus';
 import { CustomUserStatusRaw } from './CustomUserStatus';
-import LivechatAgentActivityModel from '../models/LivechatAgentActivity';
+import LivechatAgentActivityModel from '../../../app/models/server/models/LivechatAgentActivity';
 import { LivechatAgentActivityRaw } from './LivechatAgentActivity';
-import StatisticsModel from '../models/Statistics';
+import StatisticsModel from '../../../app/models/server/models/Statistics';
 import { StatisticsRaw } from './Statistics';
-import NotificationQueueModel from '../models/NotificationQueue';
+import NotificationQueueModel from '../../../app/models/server/models/NotificationQueue';
 import { NotificationQueueRaw } from './NotificationQueue';
-import LivechatBusinessHoursModel from '../models/LivechatBusinessHours';
+import LivechatBusinessHoursModel from '../../../app/models/server/models/LivechatBusinessHours';
 import { LivechatBusinessHoursRaw } from './LivechatBusinessHours';
-import ServerEventModel from '../models/ServerEvents';
+import ServerEventModel from '../../../app/models/server/models/ServerEvents';
 import { UsersSessionsRaw } from './UsersSessions';
-import UsersSessionsModel from '../models/UsersSessions';
+import UsersSessionsModel from '../../../app/models/server/models/UsersSessions';
 import { ServerEventsRaw } from './ServerEvents';
-import { trash } from '../models/_BaseDb';
-import LoginServiceConfigurationModel from '../models/LoginServiceConfiguration';
+import { trash } from '../../../app/models/server/models/_BaseDb';
+import LoginServiceConfigurationModel from '../../../app/models/server/models/LoginServiceConfiguration';
 import { LoginServiceConfigurationRaw } from './LoginServiceConfiguration';
 import { InstanceStatusRaw } from './InstanceStatus';
-import InstanceStatusModel from '../models/InstanceStatus';
+import InstanceStatusModel from '../../../app/models/server/models/InstanceStatus';
 import { IntegrationHistoryRaw } from './IntegrationHistory';
-import IntegrationHistoryModel from '../models/IntegrationHistory';
-import OmnichannelQueueModel from '../models/OmnichannelQueue';
+import IntegrationHistoryModel from '../../../app/models/server/models/IntegrationHistory';
+import OmnichannelQueueModel from '../../../app/models/server/models/OmnichannelQueue';
 import { OmnichannelQueueRaw } from './OmnichannelQueue';
-import EmailInboxModel from '../models/EmailInbox';
+import EmailInboxModel from '../../../app/models/server/models/EmailInbox';
 import { EmailInboxRaw } from './EmailInbox';
-import EmailMessageHistoryModel from '../models/EmailMessageHistory';
+import EmailMessageHistoryModel from '../../../app/models/server/models/EmailMessageHistory';
 import { EmailMessageHistoryRaw } from './EmailMessageHistory';
 import { api } from '../../sdk/api';
 import { initWatchers } from '../../modules/watchers/watchers.module';
